Updates per PR review.
