+++ src/languages/lib/ecmascript.js
@@ -1,4 +1,12 @@
 export const IDENT_RE = '[A-Za-z$_][0-9A-Za-z$_]*';
+export const EXTENDED_NUMBER_RE = '([-+]?)(\\b0[xX][a-fA-F0-9]+|(\\b\\d+(\\.\\d*)?|\\.\\d+)([eE][-+]?\\d+)?)|NaN|[-+]?Infinity'; // 0x..., 0..., decimal, float
+
+export const EXTENDED_NUMBER_MODE = {
+  scope: 'number',
+  begin: EXTENDED_NUMBER_RE,
+  relevance: 0
+};
+
 const KEYWORDS = [
   "as", // for exports
   "in",
